requirement.txt
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
APScheduler==3.6.3
argcomplete==3.1.4
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
Babel==2.10.3
bcc==0.29.1
blessings==1.7
blinker==1.7.0
Brlapi==0.8.5
cachetools==4.2.2
certifi==2024.6.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.6
cloud-init==24.4.1
colorama==0.4.6
command-not-found==0.3
configobj==5.0.8
constantly==23.10.4
cryptography==42.0.7
cupshelpers==1.0
dbus-python==1.3.2
defer==1.0.6
distlib==0.3.8
distro==1.9.0
distro-info==1.7+build1
Django==5.0.6
django-cors-headers==4.3.1
django-geoposition==0.3.0
django-redis==5.4.0
djangorestframework==3.15.1
djangorestframework_simplejwt==5.4.0
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
earthengine-api==0.1.405
ee==0.2
filelock==3.13.1
frozenlist==1.4.1
GDAL==3.8.4
google-api-core==2.19.0
google-api-python-client==2.131.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
googlemaps==4.10.0
greenlet==3.0.3
gyp==0.1
httplib2==0.22.0
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
Jinja2==3.1.2
jsonpatch==1.32
jsonpointer==2.0
jsonschema==4.10.3
language-selector==0.1
launchpadlib==1.11.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
louis==3.29.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
msgpack==1.0.3
multidict==6.0.5
netaddr==0.8.0
numpy==1.26.4
oauthlib==3.2.2
olefile==0.46
package-name==0.1
packaging==24.0
paho-mqtt==1.6.1
pexpect==4.9.0
pillow==10.3.0
pipx==1.4.3
platformdirs==4.2.0
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.8
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycairo==1.25.1
pycparser==2.22
pycups==2.0.1
Pygments==2.17.2
PyGObject==3.48.2
PyJWT==2.8.0
pynvim==0.5.0
pyOpenSSL==24.1.0
pyparsing==3.1.2
pyrsistent==0.20.0
pyserial==3.5
python-apt==2.7.7+ubuntu4
python-dateutil==2.8.2
python-debian==0.1.49+ubuntu2
python-dotenv==1.0.1
python-telegram-bot==13.13
pytz==2024.1
pyxdg==0.28
PyYAML==6.0.1
razorpay==1.4.2
redis==5.0.4
requests==2.32.3
rest-framework-simplejwt==0.0.2
retrying==1.3.4
rich==13.7.1
rsa==4.9
service-identity==24.1.0
setuptools==68.1.2
six==1.16.0
speedtest-cli==2.1.3
sqlparse==0.5.0
systemd-python==235
texttable==1.6.7
tornado==6.4
twilio==9.1.1
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.12.1
tzlocal==5.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.2
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==2.2.1
userpath==1.9.1
virtualenv==20.25.0+ds
wadllib==1.3.6
websocket-client==1.7.0
wheel==0.42.0
xdg==5
xkit==0.0.0
yarl==1.9.4
zope.interface==6.4.post2
